src/requirements.txt
lxml==4.9.2
PyYAML==6.0
requests==2.28.1
schedule==1.1.0
